Refactored interpretation.py to include all classification type analysis
Added ability to run different types of evaluations
